Refactor: GridControls.jsx Changes: - Rename GridControls to ViewControls - Separated LibrarySearchbar and GridLayoutControls as single components - Fixed bug where grid layout doesnt toggle in mobile screens
